backend/requirements.txt
aenum==3.1.15
amqp==5.2.0
APScheduler==3.10.1
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
cashfree-pg==4.2.3
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.5.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-hijack==3.6.0
django-import-export==3.3.7
djangorestframework==3.15.1
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.7
kombu==5.3.6
MarkupPy==1.14
marshmallow==3.23.2
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.2
pillow==10.2.0
prompt-toolkit==3.0.43
pydantic==1.10.17
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
sentry-sdk==1.32.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stripe==9.4.0
razorpay==1.4.2
tablib==3.5.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
